requirements.txt
requests>=2.31.0
yt-dlp>=2023.12.30
Pillow>=10.1.0
customtkinter>=5.2.0
